Fix dependency injection Remove old logging
