+++ src/XMLFormatterModel/XmlFormatterModel.csproj
@@ -6,6 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="7.0.0" />
     <ProjectReference Include="..\PluginFramework\PluginFramework.csproj" />
     <PackageReference Include="Octokit" Version="0.47.0" />
   </ItemGroup>
