+++ src/XmlFormatterOsIndependent/XmlFormatterOsIndependent.csproj
@@ -38,6 +38,7 @@
     <PackageReference Include="Markdown.Avalonia.Tight" Version="0.10.14" />
     <PackageReference Include="MessageBox.Avalonia" Version="2.1.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="7.0.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\PluginFramework\PluginFramework.csproj" />
